Added printk
This is a rather feature-rich printk, with a more verbose verb
syntax than normal. The former is because I prefer print-driven
debugging to using a debugger and the latter is because I can
never remember C's verb format for integers. Copied from the
documentation:

printk is somewhere between C's and Go's
printf functions. The format verbs are
close to Go's, but use more verbose size
indicators.

Currently supported verbs:

	%b         Binary integer
	%o         Octal integer
	%d         Decimal integer
	%x         Hexadecimal integer or buffer (lowercase)
	%X         Hexadecimal integer or buffer (uppercase)
	%c         Character
	%s         String
	%h         Hexdump buffer
	%%         Percent literal

Verb modifiers (which appear between the
percent and the verb):

	u{n}       Unsigned integer (of n bits)
	+{n}       Signed integer (of n bits)
	m{n}       Memory buffer (of n bytes)
	{ }        Separate each byte with a space
	w{n}       Minimum width (of n chars), prefixing with spaces
	0{n}       Minimum width (of n chars), prefixing with zeros

Integers

Integers are printed with a base indicated
by the verb used (base 2 for %b, base 8 for
%o, base 10 for %d, and base 16 for %x/%X).
Integers must not be accompanied by the
memory (%m) or space (% ) modifiers, although
%mh is used to print buffers. Integers may
use the unsigned (%u), signed (%+), width (%w),
and zero-prefix (%0) modifiers. The unsigned
and signed modifiers cannot be used together,
nor can the width and zero-prefix modifiers.
To avoid ambiguity, the width or zero-prefix
modifier must come before the unsigned or
signed modifier.

Common examples:

	printk("%u8d", 255);    // "255"
	printk("%+16o", -0777); // "-777"
	printk("%08u8b", 7);    // "00000111"
	printk("%w3u8d", 37);   // " 37"
	printk("%u8x", 255);    // "ff"
	printk("%u8X", 255);    // "FF"

Characters

Characters are printed using %c. This verb
takes no modifiers.

Common examples:

	printk("%c", 'a');   // "a"
	printk("%c", 97);    // "a"

Strings

Strings are printed using %s. If the memory
modifier (%m) is used, it specifies the number
of bytes to print and indicates that the string
is passed as a `char*`. If the memory modifier
is absent, a `string` is expected (not yet
implemented). The width modifier (%w) can be
used to set the minimum width of the string.
No other modifiers can be used with %s.

Common examples:

	printk("%m5s", "Hello, World");    // "Hello"
	printk("%w7m5s", "Hello, World");  // "  Hello"

Buffers

Buffers are printed with in hexadecimal format
with %x/%X or in hexdump format with %h. The
memory modifier (%m) is used to indicate a
buffer and specifies the number of bytes from
the buffer to print. The unsigned (%u), signed
(%+), width (%w), and zero-prefix (%0) modifiers
cannot be used with buffers. The space modifier
(% ) can be used with %x/%X to add a space
between each byte.

Common examples:

	printk("%m4x", "asdk");    // "6173646b"
	printk("%m4 X", "asdk");   // "61 73 64 6B"
	printk("%m41h", "C is an open source programming language.");
		"00000000  43 20 69 73 20 61 6e 20  6f 70 65 6e 20 73 6f 75  |C is an open sou|\n"
		"00000010  72 63 65 20 70 72 6f 67  72 61 6d 6d 69 6e 67 20  |rce programming |\n"
		"00000020  6c 61 6e 67 75 61 67 65  2e                       |language.|\n"

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>