Switch to Pure64 for the loader, move to 64-bit long mode
The previous loader was quite basic and operated in 32-bit
protected mode. By taking advantage of the amazing work by
ReturnInfinity in Pure64, we now have a much more sophisticated
loader, which pushes us up to 64-bit long mode.

This has also meant considerable changes to the compilation
process. The C code is all compiled for x86_64, with nasm
used to compile the assembly code. The Firefly kernel is still
100% C. Rather than just a raw kernel, we now produce a raw
disk image.