Initialise present page table counts in mem_Init
In addition to the standard paging behaviour, as defined in the Intel
manuals, Firefly uses bits 62-53 of each
PML4E and PDPTE to store the number of entries in the table referenced
that are currently present. For example, PML4[0] will contain the
address of a PDPT. Bits 62-53 of PML4[0] will show how many of the 512
entries in the PDPT are currently present. This enables the efficent
optimisation of marking the entire PDPT absent in the PML4E if no PDPT
entries are present.