Read the kernel from disk, rather than from memory
Previously we were limited to 32 KiB for both the bootloader and
the kernel, as these were copied into memory using BIOS functions.
This limited the size of the kernel to 28 KiB, which we're already
beginning to exceed.

This change instead reads the kernel binary from disk to memory,
which avoids the 32 KiB limit. So that the bootloader knows how
many bytes to copy, the image builder writes the size (as a number
of 512-byte sectors) into the last four bytes of the bootloader in
the image. The bootloader can then read that value from memory.

This uses ATA PIO mode, with logical block addresses. It's not the
most efficient mode, but it's simple and highly supported. I'd be
very open to something more efficient.