Added an integration test for handling stack overflows
Getting a stack overflow in the kernel can easily lead to a
triple fault, as the interrupt handlers responding to the
issue will push values onto the (now invalid) stack.

The GDT has to be set up to use an interrupt stack table
that has known-good stack pointers for certain interrupt
handlers.