Fixed tests, using cargo test
The various changes over the last year or so broke testing with
cargo xtest, but we can now use cargo test instead.

Note that we had to remove the (admittedly redundant) declarations
of `panic = "abort"` in both `dev` and `release` profiles in
Cargo.toml, thanks to https://github.com/rust-lang/cargo/issues/7359.

The only thing left is the repeated deprecation warning from
bootloader, tracked in https://github.com/rust-osdev/bootloader/issues/107,
which needs to be reopend at time of writing.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>