Added a bitmap frame allocator implementation
This adds a slightly more sophisticated physical memory manager.
The most important difference between this and the existing boot
info frame allocator is that the bitmap allocator can deallocate
frames.

The bitmap allocator can also repossess an existing boot info
frame allocator, taking ownership of the frames it had allocated
and marking them as in use.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>