Added a static physical memory manager to coordinate allocations
The new BitmapFrameAllocator is nice, but we need to ensure that
only one instance exists, as we only have one set of physical
memory.

The allocator is initialised as soon as the heap is ready, then
can be debugged. Once we start to allocate memory for other uses
(particularly for userspace processes), we'll have the physical
allocator to use.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>