Place the boot info at a predictable address, store in constants
Previously, we would only get the boot info from the bootloader
via the entrypoint in kernel_main. This info would be stored in
an unpredictable address. Now, we specify the address (just below
the kernel's heap), which means we can store its address in a
constant. This means we can be more precise when debugging the
PML4.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>