Replace the stored memory map with precise virtual addresses
Previously, we stored a copy of the physical memory map provided
by the bootloader. We used this to debug the PML4 by working out
which physical frames a set of virtual pages was mapped to. This
was prone to misdiagnoses and wasn't hugely efficient.

Now that we have a better stage two physical frame allocator (see
05a2590), we shouldn't need to keep hold of the memory map. At the
same time, we can replace its original purpose by using a few
pointers to identify important parts of the kernel binary.

The only remaining addresses, currently unidentified by the PML4
debugger, are those used by the bootloader, that we can unmap.
That will come next, but for now, I'm just removing the extra data.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>