Simplify memory constants by storing them as VirtAddr values
Previously, most memory constants were of type usize, which caused
some fiddly changes and lots of calls to `.as_u64()`. This is mainly
because there isn't a safe `const fn` constructor for VirtAddr.

What I've done as a slight (but fairly safe) hack is to initialise
the constants with the unsafe `try_new` constructor, then have a
`check` function to confirm that the constant really is correct
at runtime. This has simplified a load of code that uses the
constants.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>