Moved the mapping analysis code into its own module
Previously, we had functionality in the debug module to iterate
through a set of level 4 page tables, discovering contiguous
mappings and printing them as we went along. This all worked fine.

It will also be useful to iterate through the mappings for other
purposes, so I've moved the analysis code into its own module and
separated it from the printing behaviour. This means the debug
module is now nice and simple, and the mapping code returns
structured data that can be used for other purposes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>