Switched to a higher half kernel
We're now a higher half kernel, meaning that all kernel data is
now mapped in the higher half of the virtual address space,
starting at 0xffff800000000000. This makes it much easier to
differentiate kernel memory from user memory, which now ends at
0x7fffffffffff.

In the process of making the change, I've also made the mappings
more compact, while still retaining highlights to specific address
spaces. For example, addresses starting with 0xffff8000444 are
kernel heap addresses, while those starting 0xffff8000555 are
kernel stack addresses. We also reserve plenty of space for the
kernel binary to grow, up to a current limit of 1 GiB, although
that can always be changed later if needs be.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>