Added a VirtAddrRange type to simplify use of address constants
Previously, we had lots of public constants for various important
addresses, such as the start and end of the kernel's heap and stack.
This worked fine, but was often clumsy. Now, I've added a type for
a contiguous sequence of virtual addresses. This provides various
helper functions (all const) for determining the size of the range
and whether other addresses, or ranges of addresses exist within
the range. This has led to a significant simplification of the
address constants and their processing.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>