Made the remap_kernel more strict and complete
This is a small change with a big impact. Previously, remap_kernel
unmapped unidentified pages (set up and used by the bootloader but
unnecessary once the kernel is up and running) and marked the stack
as non-executable. This was all very well, but it's based on the
assumption that the stack's other flags are fine and that the other
mappings are all fine.

Now, we take a different approach, where we set all mapping flags
directly, irrespective of their previous values. We still unmap
unidentified pages (and the null page and all of userspace), but
then instead of checking the current flags and adjusting, we just
declare the flags we want for each kind of page. The groups are
as follows:

Read-write no-execute pages:
- the kernel's stack
- the kernel's heap
- the kernel's static variables
- the mapping of all physical memory

Read-only no-execute pages:
- the kernel's constants
- the kernel's string constants
- the boot info

Read-execute pages:
- the kernel's code

Note that for all of these mappings, we now set the global flag,
meaning that they should be preserved across paging, more efficently.
Since the kernel (and the other pages it uses) will always be mapped,
no matter what else we're doing, this may help us a little. Even if
it doesn't affect performance in practice, it's a good visual
reminder of the intent.

I was going to set the parent table flags as well, so we could set
the global flag there too, but it turns out the global flag has no
effect when the table entry doesn't refer to a physical frame. The
writable flag is also important to set (or it prevents us updating
the lower tables). All in all, there's little reason to set the
parent table flags to anything other than the default.

Any remaining pages (which in practice is only pages in the kernel
binary's address space where we haven't been able to identify the
specific segment) print a warning but otherwise do nothing.

The other notable difference is that since we're making many more
changes to the paging, we only flush the TLB once, after we've
made all our changes. This will be a bit more efficient than
flushing for each page as we go along.

Also note that this was made a bit more fiddly by dealing with
generic page types (given the different page sizes). I couldn't
work out how to do it neatly in one place, so it ended up with
a bit of code repetition, but it's not too bad.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>