Identify the kernel's stack guard page, shifted stack
The main two changes here are to document and identify the guard
page before the kernel's stack to help detect stack overflows,
and to shift the stack down a page so it ends on a neater address.
I've also updated the docs in a few places to be clearer.

I would have liked to update the double fault handler to identify
kernel stack overflows with more confidence, but the stack pointer
could still be in valid stack space. I may come back to this later.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>