Added basic support for memory-mapped I/O
MMIO functionality is fairly basic, but it does the trick. You
provide a set of physical memory frames and get back an MMIO
range, which tracks the virtual address space it can use. The
read and write methods can be used to read/write anything with
the Copy trait, but it'll normally be an integer type.

Everything is bounds checked, so only the initial map call is
unsafe. This is also the first example of using the kernel_pml4
function to get the mapper and pmm::ALLOCATOR to get the physical
memory allocator.

To make space for the MMIO addres space, I've moved the physical
memory offset up slightly.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>