Added the ability to allocate sequential physical frames
There are some cases where we need to be able to allocate more
than one frame, and for the full allocation to be contiguous,
such as when using MMIO.

The implementation has some false negatives, as it won't find
sequential frames that span a boundary of the 64-bit boundary.
However, it shouldn't ever return the wrong sequence of addresses.
The implementation is fairly basic, but the similarly basic
tests work, so hopefully it's fine. Improvements here would
be very welcome.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>