Disable interrupts while modifying the IRQ handler list
By disabling interrupts while we lock the IRQs list, we can be
confident that the list is always unlocked when we handle
interrupts, allowing us to lock the list, rather than trying to
lock it and having a fallback option if that fails. This means
we should avoid ignoring interrupts because of a failed attempt
to lock the list.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>