Added initial support for kernel threads
This is a fairly big change. This introduces threads of execution,
which are initially implemented only for kernel threads. The longer
term plan is to implement threads for userspace too. The plan for
Firefly is to have a hierarchical structure of threads (the unit
of scheduling) belonging to processes (the unit of programs)
belonging to workloads (the unit of applications).

This initial support is quite primitive, but it's a good starting
point. There is no support for preemption yet, so threads must
yield to the scheduler using thread::switch(). The structures for
tracking the current process and the idle thread are not yet
CPU-local, so the implementation is not safe to run in parallel.
However, as we don't yet enable more than one CPU, this is safe
in practice. We will need to add support for CPU-local data
before we can enable multiple cores.

Calling thread::init() initialises the thread module, creating
the idle thread and preparing the current thread and the scheduler.
When thread::switch is called for the first time (in kmain), the
kernel's initial context is stored into the idle thread. From that
point on, the idle thread will execute when no other threads are
runnable. For now, the idle thread exits QEMU, to simplify
development, but this will be removed once the thread functionality
is more stable.

For now, we just demonstrate the threads functionality by switching
to a simple debugging thread, returning to the idle thread, and
exiting.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>