Added basic support for CPU-local data
This is a fairly naive implementation, where we allocate a per-CPU
region from CPU_SPACE for each CPU (which is currently just the
first CPU, as we don't have multi-CPU support yet), then we store
a pointer to the beginning of that space in the GS base. To load
the per-CPU data, we simply read the address back out of the GS
base and dereference it to access the data.

This isn't as efficient as the approach being discussed in
rust-osdev/x86_64#257, but it works well enough for now.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>