Added a thread::start function to hand off to the scheduler
Previously, we handed off to the scheduler in an ad-hoc way, by
calling thread::switch to prime the scheduler and then thread::idle_loop
if we returned as the idle thread.

This moves that functionality into a diverging function in the
thread module, allowing us to make idle_loop unexported and make
any future changes in one place.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>