Simplify thread reference management while admitting its faults
The reference management for threads is a bit poor and needs to be
improved. The intended design is that each thread exists in one
base location, which is the THREADS vector for normal threads and
the CPU-local idle thread for idle threads. There may also be an
additional reference in the CPU-local current thread. Finally, there
may be a final reference if the thread is currently being modified
in a function.

Sadly, that doesn't seem to be how things are working right now.
For now at least, I've simplified the management code until it
can be fixed properly. This commit is a bit rubbish.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>