Add initial support for pre-emption
This is a little rough and ready, but it gets us started. I'll
feel more happy once the reference management for the thread
structure is more nailed down.

The implementation is straightforward. The PIT decrements the
current thread's time slice on each tick. If the thread is left
with a time slice of zero, then we return it to a full time slice
and switch to the next thread (which may be the same thread again).

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>