Use a more lightweight thread switch when a thread exits
Previously, we would store the current thread state onto its
stack and save its stack pointer before switching to the next
thread, even while the current thread is exiting. The problem
with this is that it prevents us from safely clearing up the
exiting thread's resources.

Now, we can drop the thread and immediately switch to the next
thread without using the current thread's stack.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>