Finally fix our handling of references to exiting threads.
The intent is that there are always between one and three handles
to any thread at any one time:

1. One in either THREADS or the CPU-local idle thread.
2. Maybe one in the CPU-local current thread.
3. Maybe one in a function modifying a thread.

When a thread is exiting, we check that it shouldn't be the idle
thread, so that rules out 1b. By the time we get to switching to
the new thread, we've updated the current thread, so that rules
out 2. There shouldn't be any other references to the exiting
thread, so that rules out 3.

This means there should now be only one reference to the exiting
thread just before we switch. This means we can drop it, releasing
the thread's resources, then jump to the next thread.

Previously, this wasn't working, as the exit function was failing
to release its handle and never returned. This commit fixes that
by dropping its handle before it calls switch.

Now that the implementation is more correct, we can be more
agressive in our debug assertion, checking that exactly one
reference exists. We then drop that, dropping the thread and
its resources.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>