Eliminate the risk of memory corruption when threads exit
Previously, there was a slim race condition that risked memory
corruption when a thread exited. This was because we dropped the
exiting thread, which frees the thread's stack. The stack isn't
unmapped, but there was a slim chance that another CPU might
start a new thread, taking ownership of the stack our exiting
thread just freed. If this happened, the two threads might be
using the same stack simultaneously.

The chance of the two threads using overlapping addresses in the
shared stack are even more slim, as the new thread will start at
the top of the stack, whereas the exiting thread should be some
way into its stack. Nevertheless, the risk is non-zero.

With this commit, we use inline assembly to jump to replace_stack,
rather than calling it as a Rust function. Calling the function
pushes several registers onto the stack, as described in the
System V ABI (https://www.uclibc.org/docs/psABI-x86_64.pdf, page 21).
This is where the previous memory corruption risk lay. By using a
jump instead, we avoid touching the stack at all. This should
eliminate the risk of memory corruption entirely.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>