Added basic support for timers and resuming sleeping threads
This gives us the functionality to create a timer for a thread,
and a function to process the current timers.

A timer consists of a thread id and an instant when the thread
wants to be awoken. The idea is that when a thread needs to sleep,
it will be removed from the scheduler and a timer created to wake
it back up at the appropriate time. Calling timers::process will
iterate through the current timers (which are kept in a priority
queue, sorted by soonest timers first), processing them until it
has processed all timers, or until it reaches a timer still in
the future. For each processed timer, it removes the timer from
the queue and resumes the corresponding thread. Note that processing
the timers does not switch thread and so will not immediately
switch to any woken threads.

This is fairly basic, but it gets us starting.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>