pmm: Improve the physical memory allocator's debug format
Firstly, we now start the text with a capital letter, like we do
with most other text. Secondly, we use a minimum width of 9 digits
when printing physical addresses, a minimum width of 5 digits
for the number of used/free frames, and a minimum width of 3 digits
for the amount of memory used/free. Finally, we also print the
amount of memory used.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>