mmio: Added a way to get reference access to MMIO data
Previously, we could only read or write MMIO data. This is fine
for primitive types like integers, but is less useful for a
composite type like a structure, as it only allows us to write
the entire structure.

This gives us a way to retrieve a mutable reference to the type,
which can then be used to modify individual fields on the type.
It also simplifies the implementation of read and write.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>