thread: Added create_kernel_thread to make an unstarted thread
This takes ownership of most of the implementation of start_kernel_thread,
but doesn't add the new thread to the scheduler and marks its state as
BeingCreated, rather than Runnable.

With this new function, start_kernel_thread simply becomes a call to
create_kernel_thread, followed by scheduler::resume to start the new
thread.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>