interrupts: Enable IRQs in the PIC when registering callbacks
The PIC has a mask, determining which IRQs will be ignored by the
hardware before the CPU is notified. When we register a handler
for an IRQ, we want to ensure that the handler will actually get
notified.

To ensure this, we modify the PIC's mask as necessary to enable
the newly-registered IRQ.

This wasn't necessary before, as we had only registered a handler
for IRQ 0, which is unmasked by default, on Qemu, at least.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>