interrupts: Disable all IRQs by default
Now that we unmask IRQs in the PIC when we register handlers for
them, there's no reason to have other IRQs unmasked, only to trigger
the generic handler, which does nothing. As a result, we now mask
all IRQs until they have handlers registered.

See 89563b5 for the context around unmasking IRQs.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>