virtio: Added trait to represent a Virtio transport
Virtio has a concept of a transport, which is the mechanism used
to configure the device and perform some of the interactions with
the device (besides virtqueue interactions).

The most commonly used transport is PCI (section 4.1), but MMIO is
also specified (section 4.2). We haven't implemented either yet,
but this trait allows us to abstract away the specifics of any
given transport mechanism.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>