virtio: Added the foundations of the PCI transport
The PCI transport for Virtio provides us with a way to interact
with the device, using PCI-specific methods (particularly, PCI
capabilities and MMIO using the BARs). This commit adds code to
parse those capabilities and abstract away interactions with the
device, using safe methods.

This doesn't yet implement the Transport trait, so it is not yet
functional, but it gives us the data (and foundational features)
to do so.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>