virtio: Implement the Transport trait for PCI
This allows us to abstract away the PCI-specific details of a
Virtio device negotiated using PCI.

The code includes a common theme of retrieving a reference to the
relevant data (normally the common config structure), then using
volatile reads and writes to access the relevant fields.

Note that particular care is taken to preserve the correct data
endianness. Although this is not necessary yet (as Virtio data is
little-endian, as is x86-64), it should minimise the effort
necessary to port this to other architectures later.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>