virtio: Added buffer types for use in Virtqueues
Virtqueues has a structured concept of a buffer, which is a
contiguous sequence of physical memory from the guest machine's
perspective. Buffers can either be readable or writable by the
device.

When a buffer is passed to the device in a Virtqueue (using a
descriptor), its start address and length are provided. When a
buffer is returned to the driver by the device, the driver is
also given the minimum number of bytes that have been overwritten
from the start of the buffer. This is so that the guest knows
how much data can safely be provided to less trusted processes
without being overwritten by the guest. This prevents the guest
from accidentally revealing old memory to untrusted processes,
without having to zero the buffer before passing it to the device.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>