virtio: Added structure to represent split Virtqueue descriptors
This kicks off the functionality for split Virtqueues, as described
in section 2.6. The first important structure is the descriptor. A
descriptor contains the metadata necessary to describe a buffer
when passing it to the device.

This initial support gives us a type with the same layout as the
descriptor has in memory, plus structured flags and methods for
parsing the descriptor's flags.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>