virtio: Add structure and constructor for split Virtqueue
A split Virtqueue consists of the driver and device areas (as
already implemented), along with the descriptor area. This commit
adds a struct containing the three areas and the metadata necessary
to use them, along with a constructor to allocate and configure the
Virtqueue, using the provided transport to negotiate with the
device.

The descriptor area is implemented in the descriptors slice. The
remaining metadata consists of:

- `queue_index`: This indicates which Virtqueue this is, indexed from zero. We use this to negotiate with the transport.
- `transport`: This is our handle to the transport, allowing us to communicate with the device where necessary.
- `free_list`: We use this to track which descriptors are currently free and which have been passed to the device.
- `last_used_index`: We use this to identify how many descriptors have been returned by the device but not processed yet.

We start by identifying the amount of memory required by each area.
We then allocate the physical frames, accordingly. We zero the
memory area to ensure no stale data remains, such as incorrect flags.
With the memory allocated and zeroed, we use the transport to inform
the device of the three memory areas's physical addresses. We then
enable the queue, again, using the transport.

Finally, we use our mmio module to map the driver and device areas
and construct the state we need to operate the Virtqueue.

Note that this commit also introduces a maximum number of descriptors
per Virtqueue.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>