virtio: Implement basic split Virtqueue functionality
This commit adds basic functionality for a split Virtqueue. This
allows us to send buffers using descriptors and receive used
buffers from the device. We can also manually notify the device
that buffers have been made available using MMIO and return the
number of descriptors.

The main functionality not yet supported is more nuanced notification
suppression. For example, we don't yet support the use of the
NO_NOTIFICATIONS flag to instruct the device not to notify us,
when the VIRTIO_F_EVENT_IDX feature is not negotiated. However,
we do use the VIRTIO_F_EVENT_IDX feature to a limited extent when
it is negotiated.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>