virtio: Add a basic driver implementation
This commit adds the Virtio functionality necessary to build a
device-specific driver, such as a Virtio network driver. This
only includes the functionality generic to all Virtio drivers.

Most of the driver methods delegate to the transport or the
Virtqueues. The interesting bit is the constructor, which resets
the device, then performs feature negotiation, initialises the
requested number of Virtqueues, and returns the Driver structure.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>