virtio: Added helper structs and flags for Virtio network drivers
Virtio network drivers use a common header structure, prepended to
each packet sent or received by the device. This includes the
header flags and GSO type fields, which have separate Rust types.

Finally, we also have a Config structure to represent the config
data received from the transport that is specific to the network
driver. This can be used to retrieve the MAC address and MTU of
the device, if the device indicates support for each.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>