virtio: Added a basic network driver structure
This gives us the basis of a network driver. We take ownership of
the underlying Virtio driver and keep a copy of the device's MAC
address for convenience. We also keep a vector of the buffers we
use for sending and receiving packets.

The receive buffers are only kept so we can de-allocate them all
if the driver is dropped, as we don't otherwise keep the receive
buffers. They are all sent to the device as soon as we allocate
them, and we return them to the device as soon as their contents
have been processed.

The send buffers are a little different. We only pass them to the
device to send a packet, and return them to the vector of send
buffers once the device returns them to us. This means we may have
an incomplete list of buffers when the driver is dropped.

We also store the device's MTU, which may only be an educated guess.

The reclaim_send_buffers function can be used to retrieve buffers
from the send Virtqueue once the device has sent the contained
packet and returned the descriptor. This will typically be called
by an interrupt handler in response to a notification from the
device.

This also adds smoltcp as a dependency, as this will be used in
our initial network stack.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>