virtio: Implement a receive buffer to use as a receive token
smoltcp receives packets by redeeming a receive token. We implement
that here using a RecvBuffer, which contains the physical address
of the start of the buffer, the number of bytes written to the
buffer, and a handle to the driver. When redeemed, we construct a
byte slice from the buffer, passing it to the callback.

Once the buffer has been used, we return it to the device so it
can be used to receive future packets.

Note that for now, we only support a single buffer, and therefore
an MTU of no more than 2036 bytes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>