virtio: Implement a send buffer to use as a transmit token
smoltcp sends packets by redeeming a send token. We implement
that here using a SendBuffer, which contains a handle to the
driver. When redeemed, we take a send buffer from the driver,
then use it to construct a byte slice, passing this to the
callback.

Once the buffer has been populated, we send it to the device
so the packet can be sent.

Note that for now, we only support a single buffer, and therefore
an MTU of no more than 2036 bytes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>