virtio: Implement a smoltcp Device for a Virtio driver
smoltcp manages access to a device through the Device trait. Our
Device type contains a handle to the Virtio network driver, which
it uses to fulfil the trait. Most of this comes down to describing
the driver's capabilities (which are currently none) and returning
receive and send buffers when asked.

The important implementation choice here is that the three types
used for smoltcp (drivers::virtio::network::{Device, RecvBuffer, SendBuffer})
each has a handle to the underlying driver. This handle currently
consists of a reference-counted spinlock on the driver. This allows
us to ensure safe (synchronised) access to the driver, while the
different types have different lifetimes. It would be nice if this
wasn't necessary, as it adds a lot of overhead, but it's hard to
be confident that a slimmer implementation would be safe.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>