net: Added initial support for network interfaces and polling
This adds basic support for network interfaces, which must be a
smoltcp Interface. Note that this is initially specialised to
drivers::virtio::network::Device, as the trait bounds in the
DeviceT trait make it difficult to make the interface generic.

The main detail here is that we use a DHCP socket to add IPv4
address discovery and provide a poll method. The poll method
advances the interface's state machine, sending and receiving
packets, as necessary. We also advance the DHCP state machine,
similarly.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>