virtio: Added a basic interrupt handler for Virtio network devices
This gives us the ability to register an interrupt handler for a
Virtio network device. The handler function reclaims any send
buffers that the device has returned, then polls the interface to
process any newly-received packets.

Note the importance of reading the driver's interrupt status so
that the device de-asserts the interrupt. Without that, the handler
will be called in an infinite loop.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>