virtio: Add an installer function for Virtio network drivers
This adds an installer for Virtio network drivers configured using
PCI. We configure the device, negotiating the fetaure flags we
need, record the device's MAC address, allocate the send and
receive buffers (sending the receive buffers to the device), then
prepare the network driver.

Once the network driver is ready, we construct a smoltcp Interface
around it, register this with the network stack, then register an
interrupt handler and start a companion thread.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>